feat: handle PWA install prompt dismissal
