+++ src/App.tsx
@@ -7,7 +7,7 @@ import { useScrollToTop } from "./hooks";
 import { WhitelabelProvider } from "./context/WhitelabelContext";
 import { Toaster } from "./components/ui/toaster";
 import { Toaster as SonnerToaster } from "./components/ui/sonner";
-import InstallPrompt from "./components/InstallPrompt";
+import PwaInstallButton from "./components/PwaInstallButton";
 import {
   AuthRoutes,
   DashboardRoutes,
@@ -103,7 +103,7 @@ const App = () => {
         </Suspense>
         <Toaster />
         <SonnerToaster position="top-right" />
-        <InstallPrompt />
+        <PwaInstallButton />
       </ThemeProvider>
     </WhitelabelProvider>
   );
